Added State Domain Models and Make refinements to other models
