+++ commercetools.Sdk/commercetools.Sdk.Domain/Messages/Orders/LineItemStateTransitionMessage.cs
@@ -1,5 +1,6 @@
 using System;
 using commercetools.Sdk.Domain.Orders;
+using commercetools.Sdk.Domain.States;
 
 namespace commercetools.Sdk.Domain.Messages.Orders
 {
