+++ commercetools.Sdk/commercetools.Sdk.Domain/Orders/Order.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using commercetools.Sdk.Domain.Carts;
 using commercetools.Sdk.Domain.CustomerGroups;
+using commercetools.Sdk.Domain.States;
 
 namespace commercetools.Sdk.Domain.Orders
 {
