+++ commercetools.Sdk/commercetools.Sdk.Domain/Orders/UpdateActions/TransitionStateUpdateAction.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
+using commercetools.Sdk.Domain.States;
 
 namespace commercetools.Sdk.Domain.Orders.UpdateActions
 {
