+++ commercetools.Sdk/commercetools.Sdk.Domain/Reviews/ReviewRatingStatistics.cs
@@ -1,4 +1,4 @@
-﻿namespace commercetools.Sdk.Domain
+﻿namespace commercetools.Sdk.Domain.Reviews
 {
     public class ReviewRatingStatistics
     {
