+++ commercetools.Sdk/commercetools.Sdk.Domain/States/StateRole.cs
@@ -1,0 +1,8 @@
+namespace commercetools.Sdk.Domain.States
+{
+    public enum StateRole
+    {
+        ReviewIncludedInStatistics,
+        Return
+    }
+}
