+++ commercetools.Sdk/commercetools.Sdk.Domain/States/StateType.cs
@@ -1,0 +1,11 @@
+namespace commercetools.Sdk.Domain.States
+{
+    public enum StateType
+    {
+        OrderState,
+        LineItemState,
+        ProductState,
+        ReviewState,
+        PaymentState
+    }
+}
